Stacked borrows: Strongly protected argument calls
Miri enforces that references passed to functions are valid during the entire execution of the function, even if they are not dereferenced, and even when those references are passed inside a struct.

This was not satisfied for `single_object_encoding::Reader` because the `SerializerState` held a `ref` to the `Schema`, so the property was breaking during calls to drop.

See https://github.com/someguynamedjosh/ouroboros/issues/88 for a similar issue and discussion on that topic.