Name resolution fix
Spec actually states:
References to previously defined names are as in the latter two cases above: if they contain a dot they are a fullname, if they do not contain a dot, the namespace is the namespace of the enclosing definition.

In addition, specifying an empty struct as namespace means going back to "the null namespace".