Schema serialization fix and fullnames subtelties
Resolves #10

1. Only serialize named field first time it's encountered (obviously whoops)
2. Implement namespace shenanigans to cover for the null-namespace fullname ser/de edge-case

One might think that the spec specifies that if serializing both names and references using only "fullnames" and not filling in the "namespace" properties, we end up with a valid schema.

However, as the doc specifies about references: "if they contain a dot they are a fullname, if they do not contain a dot, the namespace is the namespace of the enclosing definition", this means that if a fullname happens to not contain a dot, it will resolve differently than it prints (because it will print as only the namespace, but resolve with the parent namespace). This is a very bad property to have, but it seems to be a bit late to change that now. (And yeah [it does seem like they did realize it very late](https://issues.apache.org/jira/browse/AVRO-1295).)

In order to dodge this issue when *declaring* names, as of the previously mentioned issue the Java implementation prints out `"name": "name", "namespace": ""` in this case where namespace is empty, instead of just `"name": "namespace.name"`: https://github.com/apache/avro/blob/03c3f64b41bc67329f969e154f68137861a88b84/lang/java/avro/src/main/java/org/apache/avro/Schema.java#L835-L844 (actually they always print namespace in a separate key but that doesn't seem necessary)

However, when writing *references*, it's not clear how they make it work:
https://github.com/apache/avro/blob/03c3f64b41bc67329f969e154f68137861a88b84/lang/java/avro/src/main/java/org/apache/avro/Schema.java#L932-L940
https://github.com/apache/avro/blob/03c3f64b41bc67329f969e154f68137861a88b84/lang/java/avro/src/main/java/org/apache/avro/Schema.java#L846-L848
It seems that this implementation would print out the fullname without dot, which would then get resolved as a short name, using the outer namespace instead of null. 🤔

The implemented solution to work around this issue is to export such names as `.name`, because the Java implementation actually would parse that (despite it being seemingly disallowed in the spec):
https://github.com/apache/avro/blob/03c3f64b41bc67329f969e154f68137861a88b84/lang/java/avro/src/main/java/org/apache/avro/Schema.java#L806-L810
This may even make this implementation more correct than the official Java one on that regard, if the Java implementation allows building such schemas through the programmatic API. 😅